refactory for fix Circular dependency when bundle
